// src/shared/types/types.ts
export interface CharacterOrigin {
  name: string;
  url: string;
}

export interface CharacterLocation {
  name: string;
  url: string;
}

export interface Character {
  id: number;
  name: string;
  status: 'Alive' | 'Dead' | 'unknown';
  species: string;
  type: string;
  gender: 'Female' | 'Male' | 'Genderless' | 'unknown';
  origin: CharacterOrigin;
  location: CharacterLocation;
  image: string;
  episode: string[];
  url: string;
  created: string;
}

export interface APIInfo {
  count: number;
  pages: number;
  next: string | null;
  prev: string | null;
}

export interface APIResponse {
  info: APIInfo;
  results: Character[];
}

export interface User {
  id: string;
  username: string;
  email: string;
  name: string;
  role: 'admin' | 'user';
}